refactor: move metrics processing to duty modules
