+++ src/duty/attestation/index.ts
@@ -1,2 +1,3 @@
 export * from './attestation.module';
 export * from './attestation.service';
+export * from './attestation.metrics';
