+++ src/duty/duty.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 
-import { ConsensusProviderModule } from '../common/eth-providers';
-import { ClickhouseModule } from '../storage/clickhouse';
+import { ConsensusProviderModule } from 'common/eth-providers';
+import { ClickhouseModule } from 'storage/clickhouse';
+
 import { AttestationModule, AttestationService } from './attestation';
+import { DutyMetrics } from './duty.metrics';
 import { DutyService } from './duty.service';
 import { ProposeModule, ProposeService } from './propose';
 import { StateModule, StateService } from './state';
@@ -11,7 +13,7 @@ import { SyncModule, SyncService } from './sync';
 
 @Module({
   imports: [AttestationModule, ProposeModule, StateModule, SyncModule, SummaryModule, ConsensusProviderModule, ClickhouseModule],
-  providers: [DutyService, AttestationService, ProposeService, StateService, SyncService, SummaryService],
-  exports: [DutyService, AttestationService, ProposeService, StateService, SyncService, SummaryService],
+  providers: [DutyService, DutyMetrics, AttestationService, ProposeService, StateService, SyncService, SummaryService],
+  exports: [DutyService, DutyMetrics, AttestationService, ProposeService, StateService, SyncService, SummaryService],
 })
 export class DutyModule {}
