+++ src/duty/index.ts
@@ -1,2 +1,3 @@
 export * from './duty.module';
 export * from './duty.service';
+export * from './duty.metrics';
