+++ src/duty/propose/index.ts
@@ -1,2 +1,3 @@
 export * from './propose.module';
 export * from './propose.service';
+export * from './propose.metrics';
