+++ src/duty/state/index.ts
@@ -1,2 +1,3 @@
 export * from './state.module';
 export * from './state.service';
+export * from './state.metrics';
