+++ src/duty/summary/index.ts
@@ -1,2 +1,3 @@
 export * from './summary.module';
 export * from './summary.service';
+export * from './summary.metrics';
