+++ src/duty/summary/summary.module.ts
@@ -1,10 +1,13 @@
 import { Module } from '@nestjs/common';
 
+import { ConsensusProviderModule } from 'common/eth-providers';
+
+import { SummaryMetrics } from './summary.metrics';
 import { SummaryService } from './summary.service';
 
 @Module({
-  imports: [],
-  providers: [SummaryService],
-  exports: [SummaryService],
+  imports: [ConsensusProviderModule],
+  providers: [SummaryService, SummaryMetrics],
+  exports: [SummaryService, SummaryMetrics],
 })
 export class SummaryModule {}
