+++ src/duty/sync/index.ts
@@ -1,2 +1,3 @@
 export * from './sync.module';
 export * from './sync.service';
+export * from './sync.metrics';
