+++ src/inspector/index.ts
@@ -1,4 +1,3 @@
 export * from './inspector.module';
 export * from './inspector.service';
 export * from './processing/data-processing.service';
-export * from './processing/stats-processing.service';
